Avoid circularity in generated `toString`
For protocols and records that have lazy members, it is possible to
create an instance that has circular references. These circular
references may lead to stack overflows in `equals`, `hashCode` and
`toString`.

A check was already present to make sure that we avoid stack overflows
in `equals` and `hashCode`, but was missing in `toString`. This commit
adds the same check to `toString`.